layout/inspector/tests/test_get_all_style_sheets.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=734861
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 734861</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 734861 **/

  function runTest() {
    var utils = SpecialPowers.Cc["@mozilla.org/inspector/dom-utils;1"]
                             .getService(SpecialPowers.Ci.inIDOMUtils);
    var res = utils.getAllStyleSheets(document);

    var foundUA = false;
    for (var i = 0; i < res.length; i++) {
      if (res[i].href === "resource://gre-resources/ua.css") {
        foundUA = true;
        break;
      }
    }
    ok(foundUA, "UA sheet should be returned with all the other sheets.");
    SimpleTest.finish();
  }

  </script>
</head>
<body onload="runTest();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=734861">Mozilla Bug 734861</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
